first test

+++ run.sh
@@ -21,13 +21,14 @@ export host=\$host
 export request_uri=\$request_uri
 
 echo "Checking for fullchain.pem"
-envsubst < /etc/nginx/conf.d/default-ssl.conf.tpl > /etc/nginx/conf.d/default.conf
-# if [ ! -e "/etc/letsencrypt/live/$DOMAIN/fullchain.pem" ]; then
-#   echo "No SSL cert, enabling HTTP only..."
-#   envsubst < /etc/nginx/conf.d/default.conf.tpl > /etc/nginx/conf.d/default.conf
-# else
-#   echo "SSL cert exists, enabling HTTPs..."
-#   envsubst < /etc/nginx/conf.d/default-ssl.conf.tpl > /etc/nginx/conf.d/default.conf
-# fi
+# envsubst < /etc/nginx/conf.d/default-ssl.conf.tpl > /etc/nginx/conf.d/default.conf
+
+if [ ! -e "/etc/letsencrypt/live/$DOMAIN/fullchain.pem" ]; then
+  echo "No SSL cert, enabling HTTP only..."
+  envsubst < /etc/nginx/conf.d/default.conf.tpl > /etc/nginx/conf.d/default.conf
+else
+  echo "SSL cert exists, enabling HTTPs..."
+  envsubst < /etc/nginx/conf.d/default-ssl.conf.tpl > /etc/nginx/conf.d/default.conf
+fi
 
 nginx -g "daemon off;"
